+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `ForgetTrackedSearcher` was not retroactively applying strategies that had a `basis`.
+
+### Changed
+- One by one verification will now only verify subclasses of the given basis.
+- Verification strategies no longer ignore parent
 
 ## [3.0.0] - 2021-06-14
 ### Added